move detr model to new structure
