+++ src/development/vortex/development/utils/data/collater/detr.py
@@ -1,6 +1,6 @@
 import torch
 
-from vortex.networks.models.detection.detr import NestedTensor
+from vortex.development.networks.models.detection.detr import NestedTensor
 from easydict import EasyDict
 
 class DETRColatte:
